+++ client/apps/game/src/ui/modules/stream/event-stream.tsx
@@ -1,6 +1,6 @@
 import { NavigateToPositionIcon } from "@/ui/components/military/army-chip";
 import { ViewOnMapIcon } from "@/ui/components/military/army-management-card";
-import { ContractAddress, getAddressNameFromEntity, getPlayerAddressFromEntity, ID } from "@bibliothecadao/eternum";
+import { ContractAddress, getAddressFromEntity, getAddressNameFromEntity, ID } from "@bibliothecadao/eternum";
 import { useDojo, world } from "@bibliothecadao/react";
 import { Component, defineComponentSystem, Entity, getComponentValue, World } from "@dojoengine/recs";
 import { getEntityIdFromKeys } from "@dojoengine/utils";
@@ -53,9 +53,7 @@ export const EventStream = () => {
         ? getComponentValue(components.Owner, getEntityIdFromKeys([BigInt(entityOwner.entity_owner_id)]))
         : getComponentValue(components.Owner, getEntityIdFromKeys([BigInt(entityId)]));
 
-      const to = eventDetails[eventType].to?.(componentValue! as any, (id: ID) =>
-        getPlayerAddressFromEntity(id, components),
-      );
+      const to = eventDetails[eventType].to?.(componentValue! as any, (id: ID) => getAddressFromEntity(id, components));
       const isPersonal = to === ContractAddress(account.address);
 
       return {
