+++ client/sdk/packages/core/src/utils/index.ts
@@ -15,7 +15,7 @@ import {
 import { ResourceCostMinMax, ResourceInputs, ResourceOutputs } from "../types";
 
 export * from "./army";
-export * from "./battleSimulation";
+export * from "./battle-simulator";
 export * from "./entities";
 export * from "./leaderboard";
 export * from "./packed-data";
